fix: move select printer into command
